+++ examples/atable/default.story.vue
@@ -33,6 +33,13 @@
 			<ATable v-model:rows="resizable_2.rows" v-model:columns="resizable_2.columns" :config="resizable_2.config" />
 		</Variant>
 
+		<Variant title="filterable">
+			<ATable
+				v-model:rows="filterable_table.rows"
+				v-model:columns="filterable_table.columns"
+				:config="filterable_table.config" />
+		</Variant>
+
 		<Variant title="loading options">
 			<ATableLoading>Loading</ATableLoading>
 			<br />
@@ -126,6 +133,69 @@ const readonly_columns: TableColumn[] = [
 	},
 ]
 
+const columns_filterable: TableColumn[] = [
+	{
+		label: 'Home Page',
+		name: 'home_page',
+		type: 'Data',
+		align: 'left',
+		edit: false,
+		width: '40ch',
+		sortable: true,
+		filterable: true,
+		filterType: 'text',
+		format: (value: { title?: string; value?: any }, context) => {
+			return `<a href="${value.title}" target="_blank">${value.title} (IP: ${context.row.ip_address})</a>`
+		},
+	},
+	{
+		label: 'HTTP Method',
+		name: 'http_method',
+		type: 'Data',
+		align: 'left',
+		edit: true,
+		width: '20ch',
+		sortable: true,
+		filterable: true,
+		filterType: 'select',
+		filterOptions: [
+			{ label: 'GET', value: 'GET' },
+			{ label: 'POST', value: 'POST' },
+			{ label: 'PUT', value: 'PUT' },
+			{ label: 'DELETE', value: 'DELETE' },
+		],
+	},
+	{
+		label: 'Status',
+		name: 'status',
+		type: 'Data',
+		align: 'center',
+		edit: true,
+		width: '15ch',
+		sortable: true,
+		filterable: true,
+		filterType: 'select', // Auto-generate options from data
+	},
+	{
+		label: 'Report Date',
+		name: 'report_date',
+		type: 'component',
+		align: 'center',
+		edit: true,
+		width: '25ch',
+		sortable: true,
+		filterable: true,
+		filterType: 'dateRange',
+		modalComponent: 'DateInput',
+		format: (value: number) => {
+			const originalDate = new Date(value)
+			const currentYear = new Date().getFullYear()
+			const updatedDate = new Date(currentYear, originalDate.getMonth(), originalDate.getDate())
+			return updatedDate.toLocaleDateString('en-US')
+		},
+	},
+]
+
 const default_table = reactive({
 	rows,
 	columns,
@@ -160,6 +230,12 @@ const full_width_table = reactive({
 	rows,
 	columns,
 	config: { view: 'list', fullWidth: true },
+})
+
+const filterable_table = reactive({
+	rows,
+	columns: columns_filterable,
+	config: { view: 'list' },
 })
 </script>